refactor(galgame): split memory_reader.py into 5 sub-modules
Move types, win32 job objects, textractor paths/handle, and process
detection out of memory_reader.py into separate _* modules.
Re-export everything from memory_reader.py to preserve public API.

Pure code-move + re-export, no logic or formatting changes.
